# nestjs-project/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/videos": {
      "post": {
        "description": "Creates the video as an `uploading` draft and opens a multipart upload in storage. The client then slices the file at `part_size` and requests presigned part URLs.",
        "operationId": "VideosController_initiateUpload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Draft created and upload session opened",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitiateUploadResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "415": {
            "description": "UNSUPPORTED_VIDEO_FORMAT: mime type or extension outside the allowlist, or mismatched",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "VIDEO_TOO_LARGE: size above 10 GiB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Initiate a video upload",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{shortId}/upload/part-urls": {
      "post": {
        "description": "Returns presigned `UploadPart` URLs (1 h TTL) for the requested parts. Callable again at any time while the upload is in progress, e.g. to re-sign expired URLs.",
        "operationId": "VideosController_signPartUrls",
        "parameters": [
          {
            "name": "shortId",
            "required": true,
            "in": "path",
            "description": "The video's 11-char short id. Malformed, unknown and not-owned ids all respond 404.",
            "schema": {
              "example": "dQw4w9WgXcQ",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignPartUrlsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Presigned part URLs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartUrlsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "VIDEO_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "UPLOAD_NOT_IN_PROGRESS: the video is no longer uploading",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "INVALID_UPLOAD_PARTS: a part number is above part_count",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Presign part upload URLs",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{shortId}/upload/parts": {
      "get": {
        "description": "Lists the parts storage already holds, so a client resuming the upload sends only the missing ones.",
        "operationId": "VideosController_listUploadedParts",
        "parameters": [
          {
            "name": "shortId",
            "required": true,
            "in": "path",
            "description": "The video's 11-char short id. Malformed, unknown and not-owned ids all respond 404.",
            "schema": {
              "example": "dQw4w9WgXcQ",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Parts already uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadedPartsResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "VIDEO_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "UPLOAD_NOT_IN_PROGRESS: the video is no longer uploading",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "410": {
            "description": "UPLOAD_SESSION_EXPIRED: the multipart upload was aborted by storage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List uploaded parts",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{shortId}/upload/complete": {
      "post": {
        "description": "Assembles the parts, checks the final size, enqueues processing and moves the video to `processing`. Idempotent: replaying it on a `processing` or `ready` video returns the current state.",
        "operationId": "VideosController_completeUpload",
        "parameters": [
          {
            "name": "shortId",
            "required": true,
            "in": "path",
            "description": "The video's 11-char short id. Malformed, unknown and not-owned ids all respond 404.",
            "schema": {
              "example": "dQw4w9WgXcQ",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Video queued for processing (or its current state on replay)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "VIDEO_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "UPLOAD_NOT_IN_PROGRESS: the video has failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "410": {
            "description": "UPLOAD_SESSION_EXPIRED: the multipart upload is gone and no assembled object exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "INVALID_UPLOAD_PARTS, VIDEO_TOO_LARGE or VIDEO_SIZE_MISMATCH (the last two mark the video failed)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "PROCESSING_QUEUE_UNAVAILABLE: the video stays uploading; retrying is safe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Complete the upload",
        "tags": [
          "videos"
        ]
      }
    }
  },
  "info": {
    "title": "StreamTube API",
    "description": "API REST do StreamTube",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "InitiateUploadDto": {
        "type": "object",
        "properties": {}
      },
      "VideoResponseDto": {
        "type": "object",
        "properties": {
          "short_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{11}$",
            "example": "dQw4w9WgXcQ"
          },
          "title": {
            "type": "string",
            "example": "clip"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "processing_status": {
            "type": "string",
            "enum": [
              "uploading",
              "processing",
              "ready",
              "failed"
            ],
            "example": "uploading"
          },
          "publication_status": {
            "type": "string",
            "enum": [
              "draft"
            ],
            "example": "draft"
          },
          "failure_reason": {
            "type": "string",
            "enum": [
              "UNSUPPORTED_FORMAT",
              "PROCESSING_FAILED",
              "SOURCE_MISSING",
              "UPLOAD_REJECTED"
            ],
            "nullable": true,
            "example": null
          },
          "mime_type": {
            "type": "string",
            "example": "video/mp4"
          },
          "size_bytes": {
            "type": "number",
            "example": 1048576
          },
          "duration_seconds": {
            "type": "number",
            "nullable": true,
            "example": null
          },
          "width": {
            "type": "number",
            "nullable": true,
            "example": null
          },
          "height": {
            "type": "number",
            "nullable": true,
            "example": null
          },
          "video_codec": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "audio_codec": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "thumbnail_url": {
            "type": "string",
            "nullable": true,
            "description": "Public URL of the generated thumbnail.",
            "example": null
          },
          "processed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": null
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "short_id",
          "title",
          "description",
          "processing_status",
          "publication_status",
          "failure_reason",
          "mime_type",
          "size_bytes",
          "duration_seconds",
          "width",
          "height",
          "video_codec",
          "audio_codec",
          "thumbnail_url",
          "processed_at",
          "created_at",
          "updated_at"
        ]
      },
      "UploadSessionDto": {
        "type": "object",
        "properties": {
          "part_size": {
            "type": "number",
            "description": "Clients slice the file at exactly this size.",
            "example": 67108864
          },
          "part_count": {
            "type": "number",
            "description": "ceil(size / part_size)",
            "example": 1
          }
        },
        "required": [
          "part_size",
          "part_count"
        ]
      },
      "InitiateUploadResponseDto": {
        "type": "object",
        "properties": {
          "video": {
            "$ref": "#/components/schemas/VideoResponseDto"
          },
          "upload": {
            "$ref": "#/components/schemas/UploadSessionDto"
          }
        },
        "required": [
          "video",
          "upload"
        ]
      },
      "SignPartUrlsDto": {
        "type": "object",
        "properties": {}
      },
      "PartUrlDto": {
        "type": "object",
        "properties": {
          "part_number": {
            "type": "number",
            "example": 1
          },
          "url": {
            "type": "string",
            "description": "Presigned UploadPart URL (PUT the bytes)."
          }
        },
        "required": [
          "part_number",
          "url"
        ]
      },
      "PartUrlsResponseDto": {
        "type": "object",
        "properties": {
          "parts": {
            "description": "Ordered by part_number.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PartUrlDto"
            }
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "parts",
          "expires_at"
        ]
      },
      "UploadedPartDto": {
        "type": "object",
        "properties": {
          "part_number": {
            "type": "number",
            "example": 1
          },
          "etag": {
            "type": "string",
            "example": "\"9b2cf535f27731c974343645a3985328\""
          },
          "size": {
            "type": "number",
            "example": 67108864
          }
        },
        "required": [
          "part_number",
          "etag",
          "size"
        ]
      },
      "UploadedPartsResponseDto": {
        "type": "object",
        "properties": {
          "part_size": {
            "type": "number",
            "example": 67108864
          },
          "part_count": {
            "type": "number",
            "example": 1
          },
          "parts": {
            "description": "Parts storage already holds, ordered by part_number.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UploadedPartDto"
            }
          }
        },
        "required": [
          "part_size",
          "part_count",
          "parts"
        ]
      },
      "CompleteUploadDto": {
        "type": "object",
        "properties": {}
      },
      "ApiErrorEnvelope": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "error": {
            "type": "string",
            "example": "INVALID_CREDENTIALS"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Invalid email or password"
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      }
    }
  }
}